fixed a bug of heterogeneous media
